Be more forgiving :angel:
Now handles `gh user/repo` style input.

Not documenting this feature since I don't want to officially support
it or encourage its use, as that means handling edge cases.

The intention is to help users who copy-paste the user/repo path from
GitHub and don't want to replace "/" with a " ".