Provide user feedback when cloning a repo
Show a spinner during a clone, which can be a long running command.

Informs the user that the plugin is busy working in the background, and
to wait patiently.

Since the clone command is given as an argument to the spin command,
we test by matching against the arguments of the spin command. We ignore
options such as --quiet and --format since we are testing behaviour, not
presentation, which is likely to change.

Thanks to @jorgebucaran for writing the handy spin plugin!