Fix plugins
Ownership of the fish-spin project has moved to americanhanko.

That would normally be fine since GitHub can likely handle the redirect
if the migration was done via the project settings.

But it seems like jorgebucaran/fish-spin was renamed and used for
replay.fish, because https://github.com/jorgebucaran/fish-spin redirects
to it. A breaking change, forcing us to fix the reference.